Update accepting-a-connection.md
Added a missing line of code to close the server socket after the accept() call. It is in "[Complete Winsock Server Code](https://learn.microsoft.com/en-us/windows/win32/winsock/complete-server-code)" but not in the tutorial. This is the appopriate place to add the explanation to this line of code.